# Conflicts:
#	nts.uk/uk.com/com.ctx/bs/person/nts.uk.ctx.bs.person.dom/src/main/java/nts/uk/ctx/bs/person/dom/person/setting/init/item/PerInfoInitValueSetItem.java
#	nts.uk/uk.com/com.ctx/bs/person/nts.uk.ctx.bs.person.infra/src/main/java/repository/person/info/setting/JpaPerInfoInitValSetItem.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/optitem/KrcstCalcResultRange_.java
@@ -17,13 +17,13 @@ public class KrcstCalcResultRange_ {
 	public static volatile SingularAttribute<KrcstCalcResultRange, KrcstCalcResultRangePK> krcstCalcResultRangePK;
 
 	/** The emp appl atr. */
-	public static volatile SingularAttribute<KrcstCalcResultRange, Short> empApplAtr;
+	public static volatile SingularAttribute<KrcstCalcResultRange, Integer> empApplAtr;
 
 	/** The upper limit atr. */
-	public static volatile SingularAttribute<KrcstCalcResultRange, Short> upperLimitAtr;
+	public static volatile SingularAttribute<KrcstCalcResultRange, Integer> upperLimitAtr;
 
 	/** The lower limit atr. */
-	public static volatile SingularAttribute<KrcstCalcResultRange, Short> lowerLimitAtr;
+	public static volatile SingularAttribute<KrcstCalcResultRange, Integer> lowerLimitAtr;
 
 	/** The upper time range. */
 	public static volatile SingularAttribute<KrcstCalcResultRange, Integer> upperTimeRange;
